moved response model to screens package
