+++ src/main/java/use_cases/modify_event_use_case/EventModifierInteractor.java
@@ -1,6 +1,8 @@
 package use_cases.modify_event_use_case;
 
 
+import screens.gui_screens.EventModifierResponseModel;
+
 public class EventModifierInteractor implements EventModifierInputBoundary {
 
     final EventModifierDsGateway eventModifierDsGateway;
