+++ src/main/java/use_cases/modify_event_use_case/EventModifierPresenter.java
@@ -1,5 +1,7 @@
 package use_cases.modify_event_use_case;
 
+import screens.gui_screens.EventModifierResponseModel;
+
 public interface EventModifierPresenter {
     EventModifierResponseModel prepareSuccessView(EventModifierResponseModel event);
     EventModifierResponseModel prepareFailView(String error);
